Sepração de admins X usuários comuns. Setar alguem como admin ainda é manual

+++ app.js
@@ -150,7 +150,11 @@ app.get("/logout", (req, res)=>{
 
 app.get('/produtos',Logado,  (req, res) => {
     Produto.find().populate('categoria').sort({nome: 'desc'}).lean().then((produtos) => {
-        res.render("produtos/produtos", {produtos: produtos});
+        if(req.user.eAdmin == 1){
+            res.render("produtos/produtos", {produtos: produtos});
+        }else{
+            res.render("produtos/produtosUser", {produtos: produtos});
+        }        
     }).catch((err) => {
         req.flash("error_msg", "Houve um erro ao listar os produtos");
         res.redirect('/');
@@ -195,53 +199,79 @@ app.post("/produtos/novo",Logado,  (req, res)=>{
 })
 
 app.post("/produtos/edit",Logado, (req,res)=>{
-    Produto.findOne({_id:req.body.id}).then((produto)=>{
-        produto.nome = req.body.nome
-        produto.descricao= req.body.descricao
-        produto.categoria = req.body.categoria
-        produto.save().then(()=>{
-            req.flash("success_msg", "Sucesso na edição")
-            res.redirect("/produtos")
+    if(req.user.eAdmin == 1){
+        Produto.findOne({_id:req.body.id}).then((produto)=>{
+            produto.nome = req.body.nome
+            produto.descricao= req.body.descricao
+            produto.categoria = req.body.categoria
+            produto.save().then(()=>{
+                req.flash("success_msg", "Sucesso na edição")
+                res.redirect("/produtos")
+            }).catch((err)=>{
+                res.flash("error_msg", "merda na edição")
+                res.redirect("/produtos")
+            })
         }).catch((err)=>{
-            res.flash("error_msg", "merda na edição")
+            req.flash("error_msg", "erro ao editar")
             res.redirect("/produtos")
-        })
-    }).catch((err)=>{
-        req.flash("error_msg", "erro ao editar")
+        })        
+    }else{
+        req.flash("error_msg", "Sómente admins podem editar")
         res.redirect("/produtos")
-    })    
+    }       
+    
 })
 
+     
+
 app.get("/produtos/edit/:id",Logado, (req,res)=>{
-    Produto.findOne({_id:req.params.id}).populate('categoria').lean().then((produto)=>{
-        Categoria.find().lean().then((categorias)=>{
-            res.render('produtos/editprodutos', {categorias:categorias, produto:produto})
+    if(req.user.eAdmin == 1){
+        Produto.findOne({_id:req.params.id}).populate('categoria').lean().then((produto)=>{
+            Categoria.find().lean().then((categorias)=>{
+                res.render('produtos/editprodutos', {categorias:categorias, produto:produto})
+            }).catch((err)=>{
+                req.flash("error_msg", "erro ao listar categorias")
+                res.redirect("/admin/postagens") 
+            })
         }).catch((err)=>{
-            req.flash("error_msg", "erro ao listar categorias")
-            res.redirect("/admin/postagens") 
-        })
-    }).catch((err)=>{
-        req.flash("error_msg", "erro ao listar produto ")
+            req.flash("error_msg", "erro ao listar produto ")
+            res.redirect("/produtos")
+        })    
+    }else{
+        req.flash("error_msg", "Sómente admins podem editar")
         res.redirect("/produtos")
-    })    
+    } 
 })
 
 app.post("/produtos/deletar",Logado, (req,res)=>{   
-    Produto.deleteOne({_id:req.body.id}).then(()=>{
-        req.flash("success_msg", "deletado com sucesso")
+    if(req.user.eAdmin == 1){
+        Produto.deleteOne({_id:req.body.id}).then(()=>{
+            req.flash("success_msg", "deletado com sucesso")
+            res.redirect("/produtos")
+        }).catch((err)=>{
+            req.flash("error_msg", "erro ao deletar produto")
+            res.redirect("/produtos")
+        })    
+    }else{
+        req.flash("error_msg", "Sómente admins podem deletar")
         res.redirect("/produtos")
-    }).catch((err)=>{
-        req.flash("error_msg", "erro ao deletar produto")
-        res.redirect("/produtos")
-    })    
+    }
 })
 
 app.get('/categorias',Logado, (req, res)=>{
-    Categoria.find().sort({date:'desc'}).lean().then((categorias)=>{
-        res.render("categorias/categorias", {categorias: categorias})
-    }).catch((err)=>{
-        req.flash("error_msg", "Houve um erro ao listar as categorias")
-    })    
+    if(req.user.eAdmin == 1){
+        Categoria.find().sort({date:'desc'}).lean().then((categorias)=>{
+            res.render("categorias/categorias", {categorias: categorias})
+        }).catch((err)=>{
+            req.flash("error_msg", "Houve um erro ao listar as categorias")
+        })    
+    }else{
+        Categoria.find().sort({date:'desc'}).lean().then((categorias)=>{
+            res.render("categorias/categoriasUser", {categorias: categorias})
+        }).catch((err)=>{
+            req.flash("error_msg", "Houve um erro ao listar as categorias")
+        })    
+    }     
 })
 
 app.get('/categorias/add',Logado, (req, res)=>{
@@ -277,23 +307,45 @@ app.post("/categorias/nova",Logado,  (req, res)=>{
 })
 
 app.get('/estoques', Logado, (req, res) => {
-    Estoque.find({ retirado: 0 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
-        res.render("estoques/estoques", {estoques: estoques});
-    }).catch((err) => {
-        req.flash("error_msg", "Houve um erro ao listar os produtos");
-        console.log(err)
-        res.redirect('/');
-    });
+    if(req.user.eAdmin == 1){
+        Estoque.find({ retirado: 0 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
+            res.render("estoques/estoques", {estoques: estoques});
+        }).catch((err) => {
+            req.flash("error_msg", "Houve um erro ao listar os produtos");
+            console.log(err)
+            res.redirect('/');
+        });
+    }else{
+        Estoque.find({ retirado: 0 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
+            res.render("estoques/estoquesUser", {estoques: estoques});
+        }).catch((err) => {
+            req.flash("error_msg", "Houve um erro ao listar os produtos");
+            console.log(err)
+            res.redirect('/');
+        });
+    } 
+    
 });
 
 app.get('/retiradas',Logado,  (req, res) => {
-    Estoque.find({ retirado: 1 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
-        res.render("estoques/retiradas", {estoques: estoques});
-    }).catch((err) => {
-        req.flash("error_msg", "Houve um erro ao listar os produtos");
-        console.log(err)
-        res.redirect('/');
-    });
+    if(req.user.eAdmin == 1){
+        Estoque.find({ retirado: 1 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
+            res.render("estoques/retiradas", {estoques: estoques});
+        }).catch((err) => {
+            req.flash("error_msg", "Houve um erro ao listar os produtos");
+            console.log(err)
+            res.redirect('/');
+        });
+    }else{
+        Estoque.find({ retirado: 1 }).populate('produto').populate('recebedor').sort({dataE: 'desc'}).lean().then((estoques) => {
+            res.render("estoques/retiradasUser", {estoques: estoques});
+        }).catch((err) => {
+            req.flash("error_msg", "Houve um erro ao listar os produtos");
+            console.log(err)
+            res.redirect('/');
+        });
+    } 
+    
 });
 
 app.get("/estoques/add",Logado,  (req, res) => {
@@ -337,38 +389,48 @@ app.post("/estoques/novo",Logado,  (req, res)=>{
 })
 
 app.get("/estoques/edit/:id",Logado, (req,res)=>{
-    Estoque.findOne({_id:req.params.id}).populate('produto').lean().then((estoque)=>{
-        Produto.find().lean().then((produtos)=>{
-            res.render('estoques/editestoques', {produtos:produtos, estoque: estoque})
+    if(req.user.eAdmin == 1){
+        Estoque.findOne({_id:req.params.id}).populate('produto').lean().then((estoque)=>{
+            Produto.find().lean().then((produtos)=>{
+                res.render('estoques/editestoques', {produtos:produtos, estoque: estoque})
+            }).catch((err)=>{
+                req.flash("error_msg", "erro ao listar categorias")
+                res.redirect("/estoques") 
+            })
         }).catch((err)=>{
-            req.flash("error_msg", "erro ao listar categorias")
-            res.redirect("/estoques") 
-        })
-    }).catch((err)=>{
-        req.flash("error_msg", "esta produto não existe")
-        res.redirect("/estoques")
-    })    
+            req.flash("error_msg", "esta produto não existe")
+            res.redirect("/estoques")
+        })    
+    }else{
+        req.flash("error_msg", "Sómente admins podem editar")
+        res.redirect("/produtos")
+    } 
 })
 
 app.post("/estoques/edit",Logado, (req,res)=>{
-    Estoque.findOne({_id:req.body.id}).then((estoque)=>{
-        estoque.quantidade = req.body.quantidade
-        estoque.observacoes= req.body.observacoes
-        estoque.dataV = req.body.dataV || Date.now(); // Set to Date.now() if not provided        
-        estoque.produto= req.body.produto
-        estoque.recebedor = req.user ? req.user._id : null // ensure user is logged in
-        estoque.save().then(()=>{
-            req.flash("success_msg", "Sucesso na edição")
-            res.redirect("/estoques")
+    if(req.user.eAdmin == 1){
+        Estoque.findOne({_id:req.body.id}).then((estoque)=>{
+            estoque.quantidade = req.body.quantidade
+            estoque.observacoes= req.body.observacoes
+            estoque.dataV = req.body.dataV || Date.now(); // Set to Date.now() if not provided        
+            estoque.produto= req.body.produto
+            estoque.recebedor = req.user ? req.user._id : null // ensure user is logged in
+            estoque.save().then(()=>{
+                req.flash("success_msg", "Sucesso na edição")
+                res.redirect("/estoques")
+            }).catch((err)=>{
+                res.flash("error_msg", "merda na edição")
+                res.redirect("/estoques")
+            })
         }).catch((err)=>{
-            res.flash("error_msg", "merda na edição")
+            console.log(err)
+            req.flash("error_msg", "erro ao editar")
             res.redirect("/estoques")
-        })
-    }).catch((err)=>{
-        console.log(err)
-        req.flash("error_msg", "erro ao editar")
-        res.redirect("/estoques")
-    })    
+        })            
+    }else{
+        req.flash("error_msg", "Sómente admins podem editar")
+        res.redirect("/produtos")
+    } 
 })
 
 app.post("/retiradas/retirar",Logado, (req,res)=>{
@@ -391,13 +453,18 @@ app.post("/retiradas/retirar",Logado, (req,res)=>{
 })
 
 app.post("/estoques/deletar",Logado, (req,res)=>{   
-    Estoque.deleteOne({_id:req.body.id}).then(()=>{
-        req.flash("success_msg", "deletado com sucesso")
+    if(req.user.eAdmin == 1){
+        Estoque.deleteOne({_id:req.body.id}).then(()=>{
+            req.flash("success_msg", "deletado com sucesso")
+            res.redirect("/produtos")
+        }).catch((err)=>{
+            req.flash("error_msg", "erro ao deletar estoque")
+            res.redirect("/estoques")
+        })
+    }else{
+        req.flash("error_msg", "Sómente admins podem deletar")
         res.redirect("/produtos")
-    }).catch((err)=>{
-        req.flash("error_msg", "erro ao deletar estoque")
-        res.redirect("/estoques")
-    })    
+    } 
 })
 
 async function getQuantidadePorCategoria(retirado) {
@@ -491,6 +558,7 @@ app.get('/estatisticas', Logado, async (req, res) => {
         res.redirect('/');
     }
 });
+
 
 const PORT = process.env.PORT ||8089
 app.listen(PORT,()=>{
